requirements.txt
google-adk
litellm
mcp
mcp-server-fetch